refactor(workflows): split TriggerToggleRow into its own file
Satisfies the 'at most 2 React components per file' CI rule — TriggerSettingsCard
now exports just the card + setting row; the toggle row lives in its own file.

Entire-Checkpoint: 5dbf2457774a
