+++ apps/web/src/features/workflows/triggers/components/TriggerSettingsCard.tsx
@@ -7,7 +7,6 @@
 
 "use client";
 
-import { Switch } from "@heroui/switch";
 import type { ReactNode } from "react";
 
 export function TriggerSettingsCard({ children }: { children: ReactNode }) {
@@ -47,35 +46,6 @@ export function TriggerSettingRow({
       >
         {children}
       </div>
-    </div>
-  );
-}
-
-interface TriggerToggleRowProps {
-  label: string;
-  hint?: string;
-  isSelected: boolean;
-  onValueChange: (value: boolean) => void;
-}
-
-export function TriggerToggleRow({
-  label,
-  hint,
-  isSelected,
-  onValueChange,
-}: TriggerToggleRowProps) {
-  return (
-    <div className="flex items-center justify-between gap-4 px-4 py-3.5">
-      <div className="flex min-w-0 flex-col gap-0.5">
-        <span className="text-sm font-medium text-zinc-200">{label}</span>
-        {hint ? <span className="text-xs text-zinc-500">{hint}</span> : null}
-      </div>
-      <Switch
-        size="sm"
-        isSelected={isSelected}
-        onValueChange={onValueChange}
-        aria-label={label}
-      />
     </div>
   );
 }
