+++ apps/web/src/features/workflows/triggers/handlers/calendar.tsx
@@ -17,8 +17,8 @@ import { IntervalPicker } from "../components/IntervalPicker";
 import {
   TriggerSettingRow,
   TriggerSettingsCard,
-  TriggerToggleRow,
 } from "../components/TriggerSettingsCard";
+import { TriggerToggleRow } from "../components/TriggerToggleRow";
 import type { RegisteredHandler, TriggerSettingsProps } from "../registry";
 import type { TriggerConfig } from "../types";
